split docs
